Fix for #25:
Problem: Too many and/or too big .css files bring gulp-filter to it's knees as when collecting .css files for processing, as I found.
This makes the 'vendors' gulp task to fail silently, stopping the dependent 'dist' task from executing properly either. The latter manifests as the missing index.html file problem in the referenced issue.

Solution: Use vinyl-fs to grab a stream of the input .css files which appears to be competent enough to handle a bigger laundry too.

A much more decent solution would be to investigate and fix the problem of gulp-filter, but that's out of my scope right now unfortunately.